Update Layout logistically for new variables
Changed layout to be in a logistical order so that ETo input values are available in display for users to "reproduce" if they want to understand how the ETo values are affected by the inputs.  Grouped similar variables so people could identify daily avg vs current observations.
Due to the number of variables in the observations, some of them are cut-off at the bottom of the screen. Moved non-critical amounts to the bottom.  A discussion has been started at UDI forums to see if a Vertical Scroll bar could be added to the Admin Console.
Also added Min/Max Humidity to Forecast [already had min/max temp]