Limit for forecasts has changed, update to reflect API limits.
Aeris has changed the limit on the amount of data a basic and
contributor plan can get for forecasts to 7 days.  Expermientation
shows that setting limit= to anything over 6 now fails.  So make 6
the maximum for the Forecast Days parameter.

Also, change the query to only query for the number of days specified
by the user.